Move api.io package to styx-common module.
