+++ components/common/src/test/java/com/hotels/styx/common/io/ResourcePathMatcher.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api.io;
+package com.hotels.styx.common.io;
 
 import com.hotels.styx.api.Resource;
 import org.hamcrest.Description;
