+++ components/common/src/test/resources/com/hotels/styx/common/io/resource.txt
