+++ components/common/src/test/resources/com/hotels/styx/common/io/subdirectory/subdir-resource.txt
