+++ components/common/src/test/resources/plugins/groovy/UrlRewrite.groovy
