+++ components/common/src/test/resources/plugins/oneplugin/url-rewrite-1.0-SNAPSHOT.jar
