+++ components/proxy/src/main/java/com/hotels/styx/StyxConfig.java
@@ -19,7 +19,7 @@ import com.google.common.base.Splitter;
 import com.hotels.styx.admin.AdminServerConfig;
 import com.hotels.styx.api.Resource;
 import com.hotels.styx.api.configuration.Configuration;
-import com.hotels.styx.api.io.ResourceFactory;
+import com.hotels.styx.common.io.ResourceFactory;
 import com.hotels.styx.client.StyxHeaderConfig;
 import com.hotels.styx.infrastructure.configuration.ConfigurationParser;
 import com.hotels.styx.infrastructure.configuration.yaml.YamlConfiguration;
