+++ components/proxy/src/test/java/com/hotels/styx/admin/handlers/VersionTextHandlerTest.java
@@ -17,7 +17,7 @@ package com.hotels.styx.admin.handlers;
 
 import com.hotels.styx.api.FullHttpResponse;
 import com.hotels.styx.api.Resource;
-import com.hotels.styx.api.io.ResourceFactory;
+import com.hotels.styx.common.io.ResourceFactory;
 import com.hotels.styx.server.HttpInterceptorContext;
 import org.testng.annotations.Test;
