+++ components/server/src/main/java/com/hotels/styx/server/HttpsConnectorConfig.java
@@ -29,7 +29,7 @@ import java.util.concurrent.TimeUnit;
 import static com.google.common.base.Objects.toStringHelper;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Strings.isNullOrEmpty;
-import static com.hotels.styx.api.io.ResourceFactory.newResource;
+import static com.hotels.styx.common.io.ResourceFactory.newResource;
 
 /**
  * Https Connector configuration.
